// src/app/proto/s2c.d.ts
export namespace S2C {

    interface IMessage {
        SC_TEST_ECHO?: (S2C.ISC_TEST_ECHO|null);
// ...
        SC_GET_REWARD?: (S2C.ISC_GET_REWARD|null);
        SC_GM_CMD?: (S2C.ISC_GM_CMD|null);
        SC_UPDATE_RECHARGE?: (S2C.ISC_UPDATE_RECHARGE|null);
        SC_ROLE_OFFLINE?: (S2C.ISC_ROLE_OFFLINE|null);
    }
// ...
    interface IReward {
        type?: (number|null);
        id?: (number|null);
        value?: (number|Long|null);
    }
// ...
    interface ISC_GET_REWARD {
        reward?: (S2C.IReward[]|null);
    }
// ...
    interface IRoleSummary {
        uid?: (number|null);
// ...
    }
// ...
    interface ISC_TEST_ECHO {
        cmdId?: (number|null);
        msg?: (string|null);
    }
// ...
}